refactor: extract JwtUtil class
