+++ src/main/java/com/alkemy/ong/web/security/JwtUtil.java
@@ -1,0 +1,73 @@
+package com.alkemy.ong.web.security;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+import java.util.function.Function;
+
+@Service
+public class JwtUtil {
+
+  @Value("${security.secret}")
+  private String secret;
+
+  @Value("${security.expiration}")
+  private String expiration;
+
+  public String extractUsername(String token) {
+    return extractClaims(token, Claims::getSubject);
+  }
+
+  public Date extractExpiration(String token) {
+    return extractClaims(token, Claims::getExpiration);
+  }
+
+  public <T> T extractClaims(String token, Function<Claims, T> claimsResolver) {
+    final Claims claims = extractAllClaims(token);
+    return claimsResolver.apply(claims);
+  }
+
+  private Claims extractAllClaims(String token) {
+    return Jwts.parser().setSigningKey(secret).parseClaimsJws(token).getBody();
+  }
+
+  private Boolean isTokenExpired(String token) {
+    return extractExpiration(token).before(new Date());
+  }
+
+  public String generateToken(UserDetails userDetails) {
+    Map<String, Object> claims = new HashMap<>();
+    userDetails.getAuthorities().forEach(grantedAuthority -> {
+      claims.put("role", grantedAuthority.getAuthority());
+    });
+    return createToken(claims, userDetails.getUsername());
+  }
+
+  private String createToken(Map<String, Object> claims, String subject) {
+    return Jwts.builder()
+        .setClaims(claims)
+        .setSubject(subject)
+        .setIssuedAt(new Date(System.currentTimeMillis()))
+        .setExpiration(new Date(
+            System.currentTimeMillis() * 100 * 60 * 60 * getExpirationInHours()
+        ))
+        .signWith(SignatureAlgorithm.HS256, secret).compact();
+  }
+
+  private int getExpirationInHours() {
+    return Integer.parseInt(Objects.requireNonNull(expiration));
+  }
+
+  public Boolean validateToken(String token, UserDetails userDetails) {
+    final String username = extractUsername(token);
+    return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
+  }
+}
